Ai/requirements.txt
requests
discord.py
openai
aiofiles
